requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
requests>=2.31.0
urllib3>=1.26.18
pydantic
pydantic-core
python-dotenv
pyyaml
langchain-core
langchain-groq
langgraph
mcp
motor
crawl4ai
charset-normalizer
certifi
idna